make similar part of model and split out filteropt
